requirements.txt
requests
Pillow==3.0.0
pywin32==219